# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MakerForge API — AI-powered hardware project generation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generate:
    post:
      operationId: generateProject
      tags: [generate]
      summary: Generate a new project package from a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInput"
      responses:
        "200":
          description: Generated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "402":
          description: Insufficient credits
  /generate/{projectId}/refine:
    post:
      operationId: refineProjectSection
      tags: [generate]
      summary: Refine a specific project section with a prompt
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefineInput"
      responses:
        "200":
          description: Updated project sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSections"

  # ── TEMPLATES ─────────────────────────────────────────────────────────────
...
  /credits/checkout:
    post:
      operationId: createCheckout
      tags: [credits]
      summary: Create Stripe checkout session for Pro subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, prompt, status, isPublic, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        prompt:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, generating, ready, error]
        isPublic:
          type: boolean
        shareSlug:
          type: ["string", "null"]
        renderImageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        skillLevel:
          type: ["string", "null"]
        estimatedCost:
          type: ["number", "null"]
        estimatedTime:
          type: ["string", "null"]
        sections:
          $ref: "#/components/schemas/ProjectSections"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProjectSections:
      type: object
      properties:
        mechanical:
          $ref: "#/components/schemas/MechanicalSection"
        electronics:
          $ref: "#/components/schemas/ElectronicsSection"
        bom:
          $ref: "#/components/schemas/BomSection"
        buildGuide:
          $ref: "#/components/schemas/BuildGuideSection"
        educationPack:
          $ref: "#/components/schemas/EducationSection"
        safety:
          $ref: "#/components/schemas/SafetySection"
    MechanicalSection:
      type: object
      properties:
        openScadCode:
          type: ["string", "null"]
        printSettings:
          type: ["string", "null"]
        materials:
          type: array
          items:
            type: string
        assemblyNotes:
          type: ["string", "null"]
    ElectronicsSection:
      type: object
      properties:
        schematicDescription:
          type: ["string", "null"]
        kiCadInstructions:
          type: ["string", "null"]
        wiringDiagram:
          type: ["string", "null"]
        powerCalcs:
          type: ["string", "null"]
        components:
          type: array
          items:
            type: string
    BomSection:
      type: object
      properties:
        tiers:
          type: object
          properties:
            budget:
              type: array
              items:
                $ref: "#/components/schemas/BomItem"
            balanced:
              type: array
              items:
                $ref: "#/components/schemas/BomItem"
            premium:
              type: array
              items:
                $ref: "#/components/schemas/BomItem"
    BomItem:
      type: object
      required: [name, quantity, unit, estimatedPrice]
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        estimatedPrice:
          type: number
        affiliateUrl:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        partNumber:
          type: ["string", "null"]
        alternatives:
          type: array
          items:
            type: string
    BuildGuideSection:
      type: object
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/BuildStep"
        toolsList:
          type: array
          items:
            type: string
        estimatedTime:
          type: ["string", "null"]
        estimatedCost:
          type: ["number", "null"]
        troubleshooting:
          type: ["string", "null"]
    BuildStep:
      type: object
      required: [stepNumber, title, description]
      properties:
        stepNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        warning:
          type: ["string", "null"]
    EducationSection:
      type: object
      properties:
        lessonPlan:
          type: ["string", "null"]
        learningObjectives:
          type: array
          items:
            type: string
        ngssAlignments:
          type: array
          items:
            type: string
        worksheetMarkdown:
          type: ["string", "null"]
        reflectionPrompts:
          type: array
          items:
            type: string
        ageGroups:
          type: array
          items:
            type: string
    SafetySection:
      type: object
      properties:
        riskScore:
          type: ["integer", "null"]
        riskLevel:
          type: ["string", "null"]
          enum: ["low", "medium", "high", null]
        disclaimers:
          type: array
          items:
            type: string
        safetyNotes:
          type: array
          items:
            type: string
        environmentalImpact:
          type: ["string", "null"]
...
    GenerateInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        skillLevel:
          type: ["string", "null"]
          enum: ["beginner", "intermediate", "advanced", null]
        budget:
          type: ["string", "null"]
          enum: ["budget", "balanced", "premium", null]
        educationMode:
          type: boolean
        templateId:
          type: ["integer", "null"]
    RefineInput:
      type: object
      required: [section, prompt]
      properties:
        section:
          type: string
          enum: [mechanical, electronics, bom, buildGuide, educationPack, safety]
        prompt:
          type: string

    # ── TEMPLATES ───────────────────────────────────────────────────────────
...
    CheckoutInput:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [pro_monthly, pro_annual]
    CheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string
...