# lib/api-zod/src/generated/types/buildGuideSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MakerForge API — AI-powered hardware project generation platform
 * OpenAPI spec version: 0.1.0
 */
import type { BuildStep } from './buildStep';

export interface BuildGuideSection {
  steps?: BuildStep[];
  toolsList?: string[];
  /** @nullable */
  estimatedTime?: string | null;
  /** @nullable */
  estimatedCost?: number | null;
  /** @nullable */
  troubleshooting?: string | null;
}