# lib/api-zod/src/generated/types/electronicsSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MakerForge API — AI-powered hardware project generation platform
 * OpenAPI spec version: 0.1.0
 */

export interface ElectronicsSection {
  /** @nullable */
  schematicDescription?: string | null;
  /** @nullable */
  kiCadInstructions?: string | null;
  /** @nullable */
  wiringDiagram?: string | null;
  /** @nullable */
  powerCalcs?: string | null;
  components?: string[];
}