# lib/api-zod/src/generated/types/safetySection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MakerForge API — AI-powered hardware project generation platform
 * OpenAPI spec version: 0.1.0
 */
import type { SafetySectionRiskLevel } from './safetySectionRiskLevel';

export interface SafetySection {
  /** @nullable */
  riskScore?: number | null;
  /** @nullable */
  riskLevel?: SafetySectionRiskLevel;
  disclaimers?: string[];
  safetyNotes?: string[];
  /** @nullable */
  environmentalImpact?: string | null;
}